% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{actorID}
\alias{actorID}
\title{actorID
A function that returns the actors's ID given the actors's name}
\usage{
actorID(reh, actorName = NULL)
}
\arguments{
\item{reh}{an reh object}

\item{actorName}{other arguments}
}
\description{
actorID
A function that returns the actors's ID given the actors's name
}
