% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{dim.remify}
\alias{dim.remify}
\title{dim.remify}
\usage{
\method{dim}{remify}(x)
}
\arguments{
\item{x}{a \code{remify} object.}
}
\description{
A function that returns the dimension of the temporal network.
}
