% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{dyad.count}
\alias{dyad.count}
\title{dyad.count}
\usage{
dyad.count(
  reh,
  actor1 = NULL,
  actor2 = NULL,
  type = NULL,
  begin = NULL,
  end = NULL
)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{actor1}{input name of actor1.}

\item{actor2}{input name of actor2.}

\item{type}{input name of event type.}

\item{begin}{first time index from which to consider events when computing the output.}

\item{end}{last time index until which to consider events when computing the output.}
}
\description{
A function that returns the count of a dyad (or a subset of dyads) that is specified according to the (original) actors and type names. The method allows for counting also on a restricted sequence of events by specifying arguments \code{begin} and/or \code{end}.
}
