% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{dyad.info}
\alias{dyad.info}
\title{dyad.info
A function that returns several information about the dyad specified according to the original names}
\usage{
dyad.info(
  reh,
  actor1 = NULL,
  actor2 = NULL,
  type = NULL,
  info = NULL,
  begin = NULL,
  end = NULL
)
}
\arguments{
\item{reh}{an reh object}

\item{actor1}{name of actor1}

\item{actor2}{name of actor2}

\item{type}{name of type}

\item{info}{information to return. It can be a vector of times when the dyad occurred, or the vector of weights}

\item{begin}{first time index from which to consider events when computing the output (computing info only for a specific time window)}

\item{end}{last time index until which to consider events when computing the output (computing info only for a specific time window)}
}
\description{
dyad.info
A function that returns several information about the dyad specified according to the original names
}
