% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{dyad}
\alias{dyad}
\alias{dyad.reh}
\title{dyad}
\usage{
dyad(reh, dyadID)

\method{dyad}{reh}(reh, dyadID)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{dyadID}{a vector of one or more dyad ID's, each one ranging from 1 to D (largest risk set size)}
}
\description{
A function that given a vector of one or more dyad ID's returns the corresponding dyad composition of "actor1", "actor2" and "type" (if event types are present). The ID's to supply must range between 1 and D (largest risk set size).
}
\section{Methods (by class)}{
\itemize{
\item \code{dyad(reh)}: dyad composition in actor1, actor2 and type from dyad ID

}}
