% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getActorName}
\alias{getActorName}
\alias{getActorName.reh}
\title{getActorName}
\usage{
getActorName(reh, actorID = NULL)

\method{getActorName}{reh}(reh, actorID = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{actorID}{a vector of actor ID's. The ID value can range between \code{1} and \code{N} (number of actors in the network)}
}
\description{
A function that given a vector of actor ID's returns the corresponding vector of actor (input) names.
}
\section{Methods (by class)}{
\itemize{
\item \code{getActorName(reh)}: return actor's name from actor's ID

}}
