% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getDyad}
\alias{getDyad}
\alias{getDyad.remify}
\title{getDyad}
\usage{
getDyad(x, dyadID)

\method{getDyad}{remify}(x, dyadID)
}
\arguments{
\item{x}{a \code{remify} object.}

\item{getDyad}{a vector of one or more dyad ID's, each one ranging from 1 to D (largest risk set size).}
}
\description{
A function that given a vector of one or more dyad ID's returns the corresponding dyad composition of "actor1", "actor2" and "type" (if event types are present). The ID's to supply must range between 1 and D (largest risk set size).
}
\section{Methods (by class)}{
\itemize{
\item \code{getDyad(remify)}: return dyad composition in actor1, actor2 and type from one (or more) dyad ID

}}
