% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getDyadID}
\alias{getDyadID}
\alias{getDyadID.reh}
\title{getDyadID}
\usage{
getDyadID(reh, actor1, actor2, type)

\method{getDyadID}{reh}(reh, actor1, actor2, type)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{actor1}{name of actor1}

\item{actor2}{name of actor2}

\item{type}{name of type}
}
\description{
A function that given a vector of names as to actor1, actor2 and type returns the corresponding dyad ID. The names to supply are the original input names of the edgelist before the processing via the function \code{remify::reh()}
}
\section{Methods (by class)}{
\itemize{
\item \code{getDyadID(reh)}: return dyad's ID from dyad's composition

}}
