% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIntereventTime}
\alias{getIntereventTime}
\title{getIntereventTime}
\usage{
getIntereventTime(time, origin, ordinal)
}
\arguments{
\item{time}{first column of the edgelist (time variable)}

\item{origin}{origin time point (t_0)}

\item{ordinal}{}
}
\description{
getIntereventTime
}
