% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getName}
\alias{getName}
\title{getName

A function that returns the name of a vector of actors or types given their ID's}
\usage{
getName(reh, actorID = NULL, typeID = NULL, ...)
}
\arguments{
\item{reh}{an reh object}

\item{actorID}{other arguments}

\item{typeID}{other arguments}

\item{...}{other arguments}
}
\description{
getName

A function that returns the name of a vector of actors or types given their ID's
}
\examples{
data("randomREH")
edgelist_reh <- reh(edgelist = edgelist,
                   actors = actors,
                   types = types, 
                   directed = TRUE,
                  ordinal = FALSE,
                   origin = origin,
                   omit_dyad = omit_dyad)
getName(reh = edgelist_reh, actorID = c(0,12,16))
}
