% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getRiskset}
\alias{getRiskset}
\title{getRiskset

A function that returns the name of a vector of actors or types given their ID's}
\usage{
getRiskset(reh)
}
\arguments{
\item{reh}{an reh object}
}
\description{
getRiskset

A function that returns the name of a vector of actors or types given their ID's
}
