% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getTypeID}
\alias{getTypeID}
\alias{getTypeID.reh}
\title{getTypeID}
\usage{
getTypeID(reh, typeName = NULL)

\method{getTypeID}{reh}(reh, typeName = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{typeName}{a vector of type names. The same names in the input edgelist.}
}
\description{
A function that given a vector of type names returns the corresponding vector of ID's.
}
\section{Methods (by class)}{
\itemize{
\item \code{getTypeID(reh)}: return type's ID from type's name

}}
