% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getTypeName}
\alias{getTypeName}
\alias{getTypeName.reh}
\title{getTypeName}
\usage{
getTypeName(reh, typeID = NULL)

\method{getTypeName}{reh}(reh, typeID = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{typeID}{a vector of type ID's. The ID value can range between \code{1} and \code{C} (number of event types in the network)}
}
\description{
A function that given a vector of type ID's returns the corresponding vector of type (input) names.
}
\section{Methods (by class)}{
\itemize{
\item \code{getTypeName(reh)}: return type's name from type's ID

}}
