% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{head.reh}
\alias{head.reh}
\title{head.reh}
\usage{
\method{head}{reh}(x, ...)
}
\arguments{
\item{x}{an \code{reh} object.}

\item{...}{further arguments to be passed.}
}
\description{
A function that returns the first 6 rows (by default) of the output edgelist.
}
