% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history}
\alias{history}
\title{History dataset}
\format{
A data frame that can be used for an analysis with remverse
}
\usage{
history
}
\description{
A randomly generated relational event history dataset for testing and examples.
}
\keyword{datasets}
