% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify_durem.R
\name{is.remify_durem}
\alias{is.remify_durem}
\title{Test whether an object is a \code{remify_durem}}
\usage{
is.remify_durem(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"remify_durem"}.
}
\description{
Test whether an object is a \code{remify_durem}
}
