% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{plot.reh}
\alias{plot.reh}
\title{plot.reh}
\usage{
\method{plot}{reh}(
  x,
  which = c(1:4),
  caption = list("Title 1", "Title 2", "Title 3", "Title 4"),
  APA_style = FALSE,
  ...
)
}
\arguments{
\item{x}{is an \code{reh} object}

\item{which}{one or more numbers between 1 and 4. Explain the numbers: (1) plots this, (2) plots that, (3) plots this and (4) plot that}

\item{caption}{list of titles for each plot}

\item{APA_style}{if \code{TRUE}, theme_remify_for_APA() will be applied, if \code{FALSE} theme_remify() will be applied by default}

\item{...}{further arguments to be passed.}
}
\description{
plot visualization
}
