% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{print.remify}
\alias{print.remify}
\title{print.remify}
\usage{
\method{print}{remify}(x, ...)
}
\arguments{
\item{x}{is a \code{remify} object.}

\item{...}{other arguments.}
}
\description{
print a summary of the event history.
}
