% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rearrangeDataFrame}
\alias{rearrangeDataFrame}
\title{rearrangeDataFrame}
\usage{
rearrangeDataFrame(x, index)
}
\arguments{
\item{x}{`data.frame` object to reorder}

\item{index}{vector with the new order}
}
\value{
data.frame `x` whose columns are rearranged according to `index`
}
\description{
rearrangeDataFrame
}
