% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rearrangeList}
\alias{rearrangeList}
\title{rearrangeList}
\usage{
rearrangeList(x, index)
}
\arguments{
\item{x}{`'list` object to reorder}

\item{index}{vector with the new order}
}
\value{
list `x` rearranged according to `index`
}
\description{
rearrangeList
}
