% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{reh}
\alias{reh}
\title{Process a Relational Event History}
\usage{
reh(
  edgelist,
  actors = NULL,
  types = NULL,
  directed = TRUE,
  ordinal = FALSE,
  origin = NULL,
  omit_dyad = NULL,
  model = c("tie", "actor")
)
}
\arguments{
\item{edgelist}{an object of class \code{\link[base]{data.frame}} characterizing the relational event history sorted by 
time with columns named `time`, `actor1`, `actor2` and optionally `type` and 
`weight`.}

\item{actors}{character vector of actors' names that may be observed interacting in the network. If \code{NULL}, actor names will be taken from the input edgelist.}

\item{types}{character vector of event types that may occur in the network. If \code{NULL}, type names will be taken from the input edgelist.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{ordinal}{logical value indicating whether only the order of events matters in the model (\code{TRUE}) or also the waiting time must be considered in the model (\code{FALSE}).}

\item{origin}{time point since when events could occur (default is \code{NULL}). If it is defined, it must have the same class of the time column in the input edgelist.}

\item{omit_dyad}{list of lists. Each list refers to one risk set modification and must have two objects: a first object named `time`, that is a vector of two values defining the first and last time point of the time window where to apply the change to the risk set and a second object, named `dyad`, which is a \code{\link[base]{data.frame}} where dyads to be removed are supplied in the format \code{actor1,actor2,type} (by row). The \code{NA} value can be used to remove multiple objects from the risk set at once with one risk set modification list (see Details).}

\item{model}{can be "tie" or "actor" oriented modeling. This argument plays a fundamental role when \code{omit_dyad} is supplied. Indeed, when actor-oriented modeling, the dynamic risk set will consist of two risk sets objects (senders' and dyads' risk sets). In the tie-oriented model the function will return a dynamic risk set referred at a dyad-level.}
}
\value{
'reh' S3 object
}
\description{
A function that processes raw data and returns a 'reh' S3 object which is used as input in other functions in \code{remverse}.
}
\details{
In \code{omit_dyad}, the \code{NA} value can be used to remove multiple objects from the risk set at once with one risk set modification list. For example, to remove all events with sender equal to actor “A” add a list with two objects \code{time = c(NA, NA)} and \code{dyad = data.frame(actor1 = A, actor2 = NA, type = NA)} to the \code{omit_dyad} list.

For more details about the \code{omit_dyad} argument, inputs, outputs, attributes and methods of \code{remify::reh()}, see \code{vignette("reh")}.
}
