% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify-package.R
\docType{package}
\name{remify-package}
\alias{remify-package}
\alias{_PACKAGE}
\title{remify: Processing and transforming REH to formats suitable for the remverse packages and more}
\description{
This package aims to process, transform a relational event history to other formats suitable for the remverse and other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remify}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remify/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@tilburguniversity.edu}

Other contributors:
\itemize{
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu} [contributor]
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu} [contributor]
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu} [contributor]
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu} [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
