% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{summary.reh}
\alias{summary.reh}
\title{summary.reh}
\usage{
\method{summary}{reh}(object, ...)
}
\arguments{
\item{object}{is an \code{reh} object}

\item{...}{further arguments to be passed.}
}
\description{
A function that returns a summary of the temporal network.
}
