% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{tail.reh}
\alias{tail.reh}
\title{tail.reh}
\usage{
\method{tail}{reh}(x, ...)
}
\arguments{
\item{x}{an \code{reh} object.}

\item{...}{further arguments to be passed.}
}
\description{
A function that returns the last 6 rows (by default) of the output edgelist.
}
