% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{typeID}
\alias{typeID}
\title{typeID
A function that returns the type's ID given they type's name}
\usage{
typeID(reh, typeName = NULL)
}
\arguments{
\item{reh}{an reh object}

\item{typeName}{other arguments}
}
\description{
typeID
A function that returns the type's ID given they type's name
}
