% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{typeName}
\alias{typeName}
\title{typeName
A function that given a vector of type ID's returns the corresponding vector of type names}
\usage{
typeName(reh, typeID = NULL)
}
\arguments{
\item{reh}{an reh object}

\item{typeID}{other arguments}
}
\description{
typeName
A function that given a vector of type ID's returns the corresponding vector of type names
}
