% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{View.reh}
\alias{View.reh}
\title{View.reh}
\usage{
View.reh(x, title)
}
\arguments{
\item{x}{an \code{reh} object.}

\item{title}{a title for the window. Default title is 'x'.}
}
\description{
A function that opens a view of the output edgelist in a new window.
}
