% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{actorID}
\alias{actorID}
\alias{actorID.reh}
\title{actorID}
\usage{
actorID(reh, actorName = NULL)

\method{actorID}{reh}(reh, actorName = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{actorName}{a vector of actor names.}
}
\description{
A function that given a vector of actor names returns the corresponding vector of ID's.
}
\section{Methods (by class)}{
\itemize{
\item \code{reh}: actor's ID from actor's name
}}

