% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{actorID}
\alias{actorID}
\title{actorID}
\usage{
actorID(reh, actorName = NULL)
}
\arguments{
\item{reh}{an \code{reh} object}

\item{actorName}{a vector of actor names}
}
\description{
A function that given a vector of actor names returns the corresponding vector of ID's
}
