% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertInputREH}
\alias{convertInputREH}
\title{convertInputREH}
\usage{
convertInputREH(
  edgelist,
  actorsDictionary,
  typesDictionary,
  M,
  directed,
  omit_dyad
)
}
\arguments{
\item{edgelist}{is the input data frame with information about [time,actor1,actor2,type,weight] by row.}

\item{actorsDictionary}{dictionary of actors names}

\item{M}{number of observed relational events}

\item{directed}{boolean if the network is directed or not}

\item{omit_dyad}{list.}

\item{typesDicitonary}{dictionary of event types}
}
\value{
cube of possible combination [actor1,actor2,type]: the cell value is the column index in the rehBinary matrix
}
\description{
convertInputREH
}
