% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getActorID}
\alias{getActorID}
\alias{getActorID.reh}
\title{getActorID}
\usage{
getActorID(reh, actorName = NULL)

\method{getActorID}{reh}(reh, actorName = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{actorName}{a vector of actor names. The same names in the input edgelist.}
}
\description{
A function that given a vector of actor names returns the corresponding vector of ID's.
}
\section{Methods (by class)}{
\itemize{
\item \code{getActorID(reh)}: return actor's ID from actor's name

}}
