% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{getRiskset}
\alias{getRiskset}
\title{getRiskset}
\usage{
getRiskset(reh)
}
\arguments{
\item{reh}{an \code{reh} object}
}
\description{
A function that returns a logical matrix where by row (time point), dyads that could occur at a specific time point assume value \code{TRUE}, \code{FALSE} otherwise.
}
