% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{plot.remify}
\alias{plot.remify}
\title{plot.remify}
\usage{
\method{plot}{remify}(
  x,
  which = c(1:4),
  caption = list("Title 1", "Title 2", "Title 3", "Title 4"),
  APA_style = FALSE,
  ...
)
}
\arguments{
\item{x}{is a \code{remify} object.}

\item{which}{one or more numbers between 1 and 4. Explain the numbers: (1) plots this, (2) plots that, (3) plots this and (4) plot that.}

\item{caption}{list of titles for each plot.}

\item{APA_style}{make the plots APA-style-ready.}

\item{...}{other arguments.}
}
\description{
visual descriptive analysis of relational event network data
}
