% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{plot.remify}
\alias{plot.remify}
\title{plot.remify}
\usage{
\method{plot}{remify}(
  x,
  which = c(1:4),
  palette = "RdYlBu",
  n_intervals = 10L,
  APA_style = FALSE,
  ...
)
}
\arguments{
\item{x}{is a \code{remify} object.}

\item{which}{one or more numbers between 1 and 4. Explain the numbers: (1) plots this, (2) plots that, (3) plots this and (4) plot that.}

\item{palette}{a palette from grDevices::hcl.pals() (default is the "RdYlBu" palette).}

\item{n_intervals}{number of time intervals for time plots (default is 10).}

\item{APA_style}{make the plots APA-style-ready. (check this argument !!)}

\item{...}{other arguments. (check this argument !!)}
}
\description{
visual descriptive analysis of relational event network data.
}
