% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{print.reh}
\alias{print.reh}
\title{print.reh}
\usage{
\method{print}{reh}(x, ...)
}
\arguments{
\item{x}{is an \code{reh} object}

\item{...}{further arguments to be passed.}
}
\description{
print a summary of the event history.
}
