% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{print.remify}
\alias{print.remify}
\title{print.remify}
\usage{
\method{print}{remify}(x, ...)
}
\arguments{
\item{x}{a \code{remify} object.}

\item{...}{further arguments.}
}
\value{
displays the same information provided by the summary method.
}
\description{
print a summary of the event history.
}
\examples{
 
# processing the random network 'randomREHsmall'
library(remify)
data(randomREHsmall)
reh <- remify(edgelist = randomREHsmall$edgelist,
              model = "tie")

# printing a summary of the processed 'remify' object
print(reh)

}
