% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{reh}
\alias{reh}
\title{reh}
\usage{
reh(
  edgelist,
  covariates = list(default = NULL),
  add_actors = NULL,
  add_types = NULL,
  directed = TRUE,
  ordinal = FALSE,
  origin = NULL,
  riskset = replicate(NROW(edgelist), NaN, simplify = FALSE)
)
}
\arguments{
\item{edgelist}{a a dataframe of relational events sorted by time: [time,sender,receiver,type,weight]}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}

\item{add_actors}{vector of actors not in the network but to be considered in the analysis}

\item{add_types}{vector of types not in the network but to considered in the analysis}

\item{directed}{dyadic events directed (TRUE) or undirected (FALSE)}

\item{ordinal}{TRUE if the only the time order of events is known, FALSE if also the time value is known}

\item{origin}{starting time point (default is NULL)}

\item{riskset}{is a list of length equal to the number of events, each object a matrix with unobserved dyads (using actors string names)}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that returns a 'reh' S3 object
}
