% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{reh}
\alias{reh}
\title{reh}
\usage{
reh(
  edgelist,
  actors = NULL,
  types = NULL,
  directed = TRUE,
  ordinal = FALSE,
  origin = NULL,
  omit_dyad = NULL,
  model = c("actor", "tie")
)
}
\arguments{
\item{edgelist}{an object of class \code{"\link[base]{data.frame}"} or 
\code{"\link[base]{matrix}"} characterizing the relational event history sorted by 
time with columns 'time', 'actor1', 'actor2' and optionally 'type' and 
'weight'.}

\item{actors}{vector of actors that may be observed interacting in the network. If \code{NULL}, actor names will be drawn from the input edgelist.}

\item{types}{vector of event types that may occur in the network. If \code{NULL}, type names will be drawn from the input edgelist.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{ordinal}{logical value indicating whether only the order of events matters in the model (\code{TRUE}) or also the waiting time must be considered in the model (\code{FALSE}).}

\item{origin}{time point since which when events could occur (default is \code{NULL}). If it is defined, it must have the same class of the time column in the input edgelist.}

\item{omit_dyad}{list of lists of two elements: `time`, that is a vector of the time points which to omit dyads from, `dyad`, which is a \code{"\link[base]{data.frame}"} where dyads to be omitted are supplied.}

\item{model}{"tie" or "actor" oriented this is useful only if omit_dyad is defined. When actor-oriented, dynamic riskset will consist of two risksets (sender, and receiver riskset), in the tie-oriented only dynamic risksets will be dyad-defined.}
}
\value{
'reh' S3 object
}
\description{
A function that processes raw data and returns a 'reh' S3 object which is used as input in other functions in \code{remverse}.
}
\details{
for more details about inputs, outputs, attributes and methods of \code{remify::reh()}, see \code{vignette("reh")}.
}
