% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{reh}
\alias{reh}
\title{reh}
\usage{
reh(
  edgelist,
  covariates = list(default = NULL),
  actors = NULL,
  types = NULL,
  directed = TRUE,
  ordinal = FALSE,
  origin = NULL,
  omit_dyad = list(default = NULL)
)
}
\arguments{
\item{edgelist}{a a dataframe of relational events sorted by time: [time,sender,receiver,type,weight]}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}

\item{actors}{vector of actors not in the network but to be considered in the analysis}

\item{types}{vector of types not in the network but to considered in the analysis}

\item{directed}{dyadic events directed (TRUE) or undirected (FALSE)}

\item{ordinal}{TRUE if the only the time order of events is known, FALSE if also the time value is known}

\item{origin}{starting time point (default is NULL)}

\item{omit_dyad}{list where each element is a list of two elements: `time`, that is a vector of time points which to omit dyads from, `dyad`, which is a data.frame where dyads to omit are supplied (see more documentation about the potentials of omit_dyad in defining time varying risksets)}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that returns a 'reh' S3 object
}
