% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rehCpp}
\alias{rehCpp}
\title{rehCpp (a function for preprocessing data)}
\usage{
rehCpp(edgelist, covariates, actors, types, directed, ordinal, origin, riskset)
}
\arguments{
\item{edgelist}{is a dataframe of relational events sorted by time: [time,actor1,actor2,type,weight]}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}

\item{actors}{vector of actors not in the network but to be considered in the analysis}

\item{types}{vector of types not in the network but to considered in the analysis}

\item{directed}{dyadic events directed (TRUE) or undirected (FALSE)}

\item{ordinal}{TRUE if the only the time order of events is known, FALSE if also the time value is known}

\item{origin}{time origin value}

\item{riskset}{is a list of length equal to the number of events, each object a matrix with unobserved dyads (using actors string names)}
}
\value{
list of objects
}
\description{
rehCpp (a function for preprocessing data)
}
