% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rehCpp}
\alias{rehCpp}
\title{rehCpp (the Rcpp alias of \code{reh()})}
\usage{
rehCpp(edgelist, actors, types, directed, ordinal, origin, omit_dyad, model)
}
\arguments{
\item{edgelist}{an object of class \code{"\link[base]{data.frame}"} or 
\code{"\link[base]{matrix}"} characterizing the relational event history sorted by 
time with columns 'time', 'actor1', 'actor2' and optionally 'type' and 
'weight'.}

\item{actors}{vector of actors that may be observed interacting in the network. If \code{NULL}, actor names will be drawn from the input edgelist.}

\item{types}{vector of event types that may occur in the network. If \code{NULL}, type names will be drawn from the input edgelist.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{ordinal}{logical value indicating whether only the order of events matters in the model (\code{TRUE}) or also the waiting time must be considered in the model (\code{FALSE}).}

\item{origin}{time point since which when events could occur (default is \code{NULL}). If it is defined, it must have the same class of the time column in the input edgelist.}

\item{omit_dyad}{list of lists of two elements: `time`, that is a vector of the time points which to omit dyads from, `dyad`, which is a \code{"\link[base]{data.frame}"} where dyads to be omitted are supplied.}

\item{model}{"tie" or "actor" oriented model}
}
\value{
list of objects with processed raw data.
}
\description{
rehCpp (the Rcpp alias of \code{reh()})
}
\details{
more details can be found at the following documentation: \link[remify]{reh}.
}
