% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rehCpp}
\alias{rehCpp}
\title{rehCpp (the Rcpp alias of \code{reh()})}
\usage{
rehCpp(edgelist, actors, types, directed, ordinal, origin, omit_dyad)
}
\description{
rehCpp (the Rcpp alias of \code{reh()})
}
\details{
more details can be found at the following documentation: \link[remify]{reh}.
}
