% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R, R/zzz.R
\docType{package}
\name{remify}
\alias{remify}
\title{remify}
\usage{
remify(input)
}
\arguments{
\item{input}{an input argument for the function \code{remify()}.}
}
\value{
otuput of \code{remify()}.
}
\description{
A function that transforms a \code{reh} object into one of the possible formats that suit external packages, or vice versa.

The remify package provides two main functions: \code{reh()} for pre-processing relational event sequences and creating a new data structure (\code{reh}) used by other packages, and \code{remify()} for transforming REH from other formats or other sources to an \code{reh} structure.
}
\section{\code{remify} functions}{

The remify functions ...
}

