% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{remify}
\alias{remify}
\title{remify: a package for processing Relational Event History (REH) data}
\description{
The remify package provides two main functions: \code{reh()} for pre-processing relational event sequences and creating a new data structure (\code{reh}) used by other packages, and \code{rehshape()} for transforming REH from other formats or other sources to an \code{reh} structure and vice versa.
}
\section{\code{remify} functions}{

The remify functions ...
}

