% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R, R/zzz.R
\docType{package}
\name{remify}
\alias{remify}
\title{remify}
\usage{
remify(input)
}
\arguments{
\item{input}{an input argument for the function \code{remify()}.}
}
\value{
otuput of \code{remify()}.
}
\description{
A function that transforms a \code{reh} object into one of the possible formats that suits external packages.

The remify package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remify} functions}{

The remify functions ...
}

