% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R, R/zzz.R
\docType{package}
\name{remify}
\alias{remify}
\title{remify}
\usage{
remify(input)
}
\arguments{
\item{input}{}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that transform the REH input in one of the possible REH formats.

The remify package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remify} functions}{

The remify functions ...
}

