% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{typeID}
\alias{typeID}
\title{typeID}
\usage{
typeID(reh, typeName = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{typeName}{a vector of type names.}
}
\description{
A function that given a vector of type names returns the corresponding vector of ID's.
}
