% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remify.R
\name{typeName}
\alias{typeName}
\alias{typeName.reh}
\title{typeName}
\usage{
typeName(reh, typeID = NULL)

\method{typeName}{reh}(reh, typeID = NULL)
}
\arguments{
\item{reh}{an \code{reh} object.}

\item{typeID}{a vector of type ID's.}
}
\description{
A function that given a vector of type ID's returns the corresponding vector of type (input) names.
}
\section{Methods (by class)}{
\itemize{
\item \code{typeName(reh)}: type's name from type's ID

}}
